Refactor calculating recent activity
